finished add member function, saving in local storage
